requirements.txt

discord.py
python-dotenv 
openai 
apscheduler<4.0.0
pytz
streamlit 
plotly 
pandas
langchain
langgraph
langchain-community
langchain-core
